E1A1 Mass and CoM of a circle with 2 square holes
Having being given a circle of radius 3 with two square holes, one in the center with a side of 2 and ne in (-2,0) with a side of i had to calculate mass and the centre f mass using Monte Carlo.The density changes as such d=1-r^2. This should be okay.